XML and Defines
Added the ability to set the defines in an xml file, for modding purposes.
Moved defines to static class and made them no longer final.

+++ src/civilisationClicker/MathFunctions.java
@@ -58,7 +58,7 @@ public class MathFunctions {
 		}
 		return parsedDouble;
 	}
-	public static double parseDouble(String toParse, int defaultOutcome) {
+	public static double parseDouble(String toParse, double defaultOutcome) {
 		double parsedDouble = 0;
 		try {
 			parsedDouble = Double.parseDouble(toParse);
